decomp: update progress metrics and enhance function matching
- Increased the count of matched functions to 455/633, improving decompiled C metrics to 71.4% functions and 36.6% bytes.
- Adjusted the decompilation queue, with 178 functions still needing semantic C and 178 readable Thumb functions remaining.
- Updated documentation and progress files to reflect the latest decompilation status, including changes in function parameters and matching accuracy.

+++ src/decompiled/sub_0802BC14.c
@@ -1,7 +1,7 @@
 #include "global.h"
 
 // @ 0x0802bc14
-s32 sub_0802BC14(u16 a)
+s32 sub_0802BC14(u32 a)
 {
     s32 count;
     s32 limit;
